Dev to Main

+++ DataModels/VM/AircraftEquipment/AircraftEquipmentDataVM.cs
@@ -10,7 +10,7 @@ namespace DataModels.VM.AircraftEquipment
 
         public string Status { get; set; }
 
-        public int AirCraftId { get; set; }
+        public long AirCraftId { get; set; }
 
         public int ClassificationId { get; set; }
